internal/action: move Labels and Milestones actions to a single file
IMHO there isn't a good reason why only Labels and Milestones had their
own action files while other objects like Issues didn't. This patch
moves both Labels and Milestones actions to the action.go generic file.
